# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DPOLens",
    "summary": "Policies and law, with citations a client can check",
    "description": "\nThe only way into DPOLens. Every surface, including the MCP server and the\ndashboard, is a client of this API.\n\nInterfaces change without notice until 1.0. Errors are\n[RFC 9457](https://www.rfc-editor.org/info/rfc9457/) problem details, and `type`\nis the field to branch on.\n",
    "version": "0.0.0"
  },
  "paths": {
...
    "/v1/search": {
      "post": {
        "tags": [
          "corpus"
        ],
        "summary": "Find the clauses that answer a question",
        "operationId": "run_search_v1_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Results"
                }
              }
            }
          },
          "401": {
            "description": "No usable credential",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "The credential may not do this",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "429": {
            "description": "The token has spent its quota for the minute",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/clauses/{key}": {
      "get": {
        "tags": [
          "corpus"
        ],
        "summary": "Read one clause by its canonical key",
        "operationId": "read_clause_v1_clauses__key__get",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "A canonical key, such as gdpr:art-17:para-1",
              "title": "Key"
            },
            "description": "A canonical key, such as gdpr:art-17:para-1"
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 20
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lang"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClauseInContext"
                }
              }
            }
          },
          "401": {
            "description": "No usable credential",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "The credential may not do this",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "429": {
            "description": "The token has spent its quota for the minute",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "404": {
            "description": "No such clause or document",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/clauses/{key}/subtree": {
      "get": {
        "tags": [
          "corpus"
        ],
        "summary": "Read a clause and everything beneath it",
        "description": "What an assistant wants once it has the top hit: the whole article, in order.",
        "operationId": "read_subtree_v1_clauses__key__subtree_get",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "A canonical key, such as gdpr:art-17",
              "title": "Key"
            },
            "description": "A canonical key, such as gdpr:art-17"
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 20
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lang"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Clause"
                  },
                  "title": "Response Read Subtree V1 Clauses  Key  Subtree Get"
                }
              }
            }
          },
          "401": {
            "description": "No usable credential",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "The credential may not do this",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "429": {
            "description": "The token has spent its quota for the minute",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "404": {
            "description": "No such clause or document",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/documents": {
      "get": {
        "tags": [
          "corpus"
        ],
        "summary": "List the documents in force",
        "operationId": "list_corpus_v1_documents_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Documents"
                }
              }
            }
          },
          "401": {
            "description": "No usable credential",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "The credential may not do this",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "429": {
            "description": "The token has spent its quota for the minute",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/documents/{slug}": {
      "get": {
        "tags": [
          "corpus"
        ],
        "summary": "One document and its top-level structure",
        "operationId": "read_document_v1_documents__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The document's slug, such as gdpr",
              "title": "Slug"
            },
            "description": "The document's slug, such as gdpr"
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentInDetail"
                }
              }
            }
          },
          "401": {
            "description": "No usable credential",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "The credential may not do this",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "429": {
            "description": "The token has spent its quota for the minute",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "404": {
            "description": "No such clause or document",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Clause": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key",
            "description": "The canonical key a citation points at"
          },
          "clause_type": {
            "type": "string",
            "title": "Clause Type"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label",
            "description": "The numbering the law itself uses, such as 1. or (a)"
          },
          "heading": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heading"
          },
          "text": {
            "type": "string",
            "title": "Text",
            "description": "The clause as it stands, verbatim"
          },
          "lang": {
            "type": "string",
            "title": "Lang"
          },
          "is_authoritative": {
            "type": "boolean",
            "title": "Is Authoritative",
            "description": "Whether this language is the one that prevails in law"
          },
          "is_normative": {
            "type": "boolean",
            "title": "Is Normative",
            "description": "False for text that explains without obliging, such as a recital"
          },
          "depth": {
            "type": "integer",
            "title": "Depth"
          },
          "document_title": {
            "type": "string",
            "title": "Document Title"
          },
          "document_slug": {
            "type": "string",
            "title": "Document Slug"
          },
          "version_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version Label"
          },
          "effective_date": {
            "type": "string",
            "format": "date",
            "title": "Effective Date"
          },
          "pack_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pack Slug",
            "description": "Null for an organisation's own policy"
          },
          "jurisdiction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jurisdiction"
          },
          "trust_tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trust Tier",
            "description": "verified when a named maintainer checked the pack against the official source, community when it has passed validation only, null for a policy"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          }
        },
        "type": "object",
        "required": [
          "key",
          "clause_type",
          "label",
          "heading",
          "text",
          "lang",
          "is_authoritative",
          "is_normative",
          "depth",
          "document_title",
          "document_slug",
          "version_label",
          "effective_date",
          "pack_slug",
          "jurisdiction",
          "trust_tier",
          "source_url"
        ],
        "title": "Clause"
      },
      "ClauseInContext": {
        "properties": {
          "clause": {
            "$ref": "#/components/schemas/Clause"
          },
          "breadcrumb": {
            "items": {
              "$ref": "#/components/schemas/Clause"
            },
            "type": "array",
            "title": "Breadcrumb"
          },
          "children": {
            "items": {
              "$ref": "#/components/schemas/Clause"
            },
            "type": "array",
            "title": "Children"
          },
          "cross_references": {
            "items": {
              "$ref": "#/components/schemas/CrossReference"
            },
            "type": "array",
            "title": "Cross References"
          }
        },
        "type": "object",
        "required": [
          "clause",
          "breadcrumb",
          "children",
          "cross_references"
        ],
        "title": "ClauseInContext"
      },
...
      "DocumentInDetail": {
        "properties": {
          "document": {
            "$ref": "#/components/schemas/Document"
          },
          "outline": {
            "items": {
              "$ref": "#/components/schemas/OutlineEntry"
            },
            "type": "array",
            "title": "Outline",
            "description": "The top level only, without text. Read further with a clause key"
          }
        },
        "type": "object",
        "required": [
          "document",
          "outline"
        ],
        "title": "DocumentInDetail"
      },
      "Documents": {
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/Document"
            },
            "type": "array",
            "title": "Documents"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "How many documents are in force, ignoring limit and offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "documents",
          "total",
          "limit",
          "offset"
        ],
        "title": "Documents"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProblemDocument": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Stable identifier for this kind of failure"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "instance": {
            "type": "string",
            "title": "Instance",
            "description": "The path that produced it"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance"
        ],
        "title": "ProblemDocument",
        "description": "The error shape, declared so the specification carries it.\n\nA client branches on `type`. `title` and `status` say what class of failure it\nis, and `detail` is written for whoever has to fix it. Some problems add a\nfield of their own, such as the permission a token is missing."
      },
...
      "Results": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/Result"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "results"
        ],
        "title": "Results"
      },
      "SearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Query",
            "description": "The question, in words"
          },
          "limit": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 10
          },
          "lang": {
            "type": "string",
            "maxLength": 20,
            "title": "Lang",
            "default": "en"
          },
          "expand": {
            "type": "string",
            "enum": [
              "none",
              "siblings",
              "parent"
            ],
            "title": "Expand",
            "description": "Also return the siblings, or the parent, of each hit",
            "default": "none"
          },
          "as_of": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of",
            "description": "Read the corpus as it stood on this date, to cite what was in force then"
          },
          "include_explanatory": {
            "type": "boolean",
            "title": "Include Explanatory",
            "description": "Include text that explains without obliging, such as a recital. Off by default, because such text reads more like a question than the article that creates the duty",
            "default": false
          },
          "explain": {
            "type": "boolean",
            "title": "Explain",
            "description": "Include each retriever's ranking, for debugging",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "SearchRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}